Renaming double-quotes file
